Resolvation of the "no techs/upgrades working in visual 2008 binaries"
The reason was, that UnitID,techID,upgradeID enumerations were of type u16,u8,u8, but %X format of sscanf expects int (32 bits or more on later platforms), and as I provided smaller type, unexpected things could happen, as vs 2005 binaries are ofcourse different the random memory access was harmfull, but in vs 2008 binaries it broke the i iterator i in the for each syntax.

The resolvation is to load standard int (32bit) using the sscanf and then just convert that int to the enum value.

git-svn-id: https://bwapi.googlecode.com/svn/trunk@403 273ecf2c-f24b-0410-b6a7-e9298bb4873f