Bug fixes: Fixed Unit::getTransport so that it works for Protoss Interceptors Fixed Unit::isLoaded so that it returns true if and only if Unit::getTransport!=NULL. Fixed Unit::getLoadedUnits so that it also returns the Interceptors currently loaded inside a Protoss Carrier. Fixed a bug where Unit::getPosition and Unit::getTilePosition would return incorrect positions for loaded units. This fixes issue 225.
Changed Unit::getLoadedUnits so that it returns a set rather than a list so that it is consistent with Unit::getInterceptors and Unit::getLarva.

Added the following functions:

Unit::getCarrier - tested and works
Unit::getInterceptors - tested and sort of works (only returns interceptors that have left the carrier at least once before). Note that this function returns all interceptors controlled by a Carrier, while getLoadedUnits will only return the interceptors currently inside the carrier. The set interceptors currently outside of the carrier is the difference of these two sets.
Unit::getNydusExit - UNTESTED


git-svn-id: https://bwapi.googlecode.com/svn/trunk@2473 273ecf2c-f24b-0410-b6a7-e9298bb4873f