EventTest now runs without error both as a normal dll and as a client process (i.e. loaded into the AIModuleLoader). Also fixed some things to prevent multiple MatchEnd events from being generated - looks like the hook for GameImpl::onGameEnd is called multiple times at the end of a match, GameImpl::onMenuFrame is called between these multiple calls to onGameEnd, and GameImpl::update is called after onMenuFrame between onGameEnd events, which all chronologically looks something like this at the end of the match:
GameImpl::onGameEnd
GameImpl::onMenuFrame (called 1 or more times)
GameImpl::update (called 1 or more times)
GameImpl::onGameEnd
GameImpl::onMenuFrame (called 1 or more times)
GameImpl::update (called 1 or more times)
GameImpl::onGameEnd

To prevent BWAPI from incorrectly recognizing these sequences of GameImpl callbacks as actual matches, onMenuFrame waits until 200 ms after the first call to onGameEnd before resetting these variables. This is not a perfect fix and a really fast computer might be able to break it by starting a new match within 200 ms after the last match end, but it appears to work correctly for now.

Ideally though it would be better to find a new hook for the onGameEnd callback so that it occurs only once and occurs after the GameImpl::update() hook is called for the last time at the end of a match.

git-svn-id: https://bwapi.googlecode.com/svn/trunk@2672 273ecf2c-f24b-0410-b6a7-e9298bb4873f