Address an issue that reduces APM by 1/3. Please tell if it produces any UI bugs.
The issue was: When the user had a unit selected, and BWAPI performed an action, the unit would be reselected after every action.
The fix: Only reselects the unit when an action is given.

How I got 1/3: If BWAPI takes two actions to order a unit (select, move), then the third action would have reselected your unit.


git-svn-id: https://bwapi.googlecode.com/svn/trunk@2743 273ecf2c-f24b-0410-b6a7-e9298bb4873f