Make Unit finder Y offset relative to X. Use our #defines instead of plain numbers. A function description for SNP module. Added UnitType::isMineralField(). Added operator int(), removed comparison operators, for all *Type objects. Avoid some casting(UnitImpl* for example) where possible. Add some spacing, remove some NULLs, etc. Some more optimizations and restrictions for unit finder.
For the most part this commit reduces the amount of code necessary for many things. I know that at least 1.5 KB of code has been removed.

If you run into any problems with this commit specifically, then please provide a detailed explanation of the problem. I will be leaving this Friday(Tomorrow) and won't be available until Monday/Tuesday, so leave an issue or email.


git-svn-id: https://bwapi.googlecode.com/svn/trunk@3483 273ecf2c-f24b-0410-b6a7-e9298bb4873f