- Fixed the original problem that was reported for Issue 373
(i.e. "auto_menu attempts to start before map transfer is processed/confirmed").  FYI, it was starting the game without waiting for the map transfer to be processed/confirmed (causing a crash for me if a player hasn't downloaded the map) when either there are no free slots in the lobby, or wait_for_time is reached and the number of players in the lobby >= wait_for_min_players.
- Fixed the "The scenario map file is still being distributed" BW error message that was displayed on the screen when using auto-menu (auto-menu was getting stuck; the error message was displayed momentarily each time I tried to start the game by manually clicking the "Ok" button).  This change is an additional fix on top of r4375 "Fixed a bug that prevented auto-menu from working correctly.".  This BW error message was reported as part of Issue 373 and part of Issue 443.  The original problem for Issue 443 (i.e. "Second client downloads maps it already has") is not fixed though - it happens for me for all maps except the ones directly under the "maps" folder.