Fixed the "Second client downloads maps it already has" part of Issue 443
the root cause was the lastMapGen string being used in Detours.cpp:_FindFirstFile() when joining a game (because it should only be used when creating a game).   The "auto-menu instability" part of Issue 443 is still occurring in some cases, e.g. crashes in some cases if you quickly open multiple multiplayer instances of BWAPI and there are more of them than the wait_for_max_players setting.  Another problem is that multiple games may be created (rather than having a single BWAPI instance as the game host and the rest of the BWAPI instances joining it), due to a race condition.