Fix client desync when trying to expand to base with blocking minerals
isVisible() is not synced to a client when a unit is removed, but it is marked as nonexistent.
Clients should always check exists() on units, not doing so makes this mineral check algorithm see a ghost mineral field. This does not occur on modules since there isVisible() is properly updated.